replace camera, compose, loading fragment with activities, changed compose feature to a FAB in home feed
