+++ .idea/misc.xml
@@ -18,8 +18,11 @@
         <entry key="app/src/main/res/drawable/thumbs_down.xml" value="0.1" />
         <entry key="app/src/main/res/drawable/thumbs_up.xml" value="0.1" />
         <entry key="app/src/main/res/drawable/translucent_overlay.xml" value="0.3735" />
+        <entry key="app/src/main/res/layout/activity_camera.xml" value="0.3029891304347826" />
         <entry key="app/src/main/res/layout/activity_connect_spotify.xml" value="0.1" />
         <entry key="app/src/main/res/layout/activity_edit_profile.xml" value="0.3428442028985507" />
+        <entry key="app/src/main/res/layout/activity_image_results.xml" value="0.3029891304347826" />
+        <entry key="app/src/main/res/layout/activity_loading.xml" value="0.3029891304347826" />
         <entry key="app/src/main/res/layout/activity_login.xml" value="0.2568585603133493" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.24764251708984375" />
         <entry key="app/src/main/res/layout/activity_signup.xml" value="0.24320652173913043" />
@@ -43,6 +46,7 @@
         <entry key="app/src/main/res/layout/item_profile.xml" value="0.3196461995442708" />
         <entry key="app/src/main/res/layout/item_song_card.xml" value="0.3328804347826087" />
         <entry key="app/src/main/res/menu/menu_bottom_navigation.xml" value="0.36944444444444446" />
+        <entry key="app/src/main/res/menu/menu_compose.xml" value="0.32760416666666664" />
         <entry key="app/src/main/res/menu/menu_editprofile_toolbar.xml" value="0.40260416666666665" />
         <entry key="app/src/main/res/menu/menu_profile_toolbar.xml" value="0.29583333333333334" />
       </map>
