+++ app/src/main/AndroidManifest.xml
@@ -46,6 +46,15 @@
             android:name=".activities.MainActivity"
             android:exported="true" />
         <activity
+            android:name=".activities.CameraActivity"
+            android:exported="false" />
+        <activity
+            android:name=".activities.ImageResultsActivity"
+            android:exported="false" />
+        <activity
+            android:name=".activities.LoadingActivity"
+            android:exported="false" />
+        <activity
             android:name=".activities.SongRecommendationsActivity"
             android:exported="true" />
         <activity
