+++ app/src/main/java/com/example/photofy/activities/LoadingActivity.java
@@ -1,0 +1,19 @@
+package com.example.photofy.activities;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+import android.util.Log;
+
+import com.example.photofy.R;
+
+public class LoadingActivity extends AppCompatActivity {
+
+    public static final String TAG = "LoadingActivity";
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_loading);
+    }
+}
