+++ app/src/main/java/com/example/photofy/activities/MainActivity.java
@@ -7,12 +7,9 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.navigation.NavController;
 import androidx.navigation.NavDestination;
-import androidx.navigation.Navigation;
 import androidx.navigation.fragment.NavHostFragment;
-import androidx.navigation.ui.NavigationUI;
 
 import android.content.Context;
-import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.util.Log;
@@ -20,10 +17,9 @@ import android.view.MenuItem;
 import android.view.View;
 
 import com.example.photofy.R;
-import com.example.photofy.fragments.CommentsFragment;
-import com.example.photofy.fragments.ComposeFragment;
 import com.example.photofy.fragments.HomeFragment;
 import com.example.photofy.fragments.ProfileFragment;
+import com.example.photofy.fragments.SearchFragment;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationBarView;
 
@@ -84,8 +80,8 @@ public class MainActivity extends AppCompatActivity {
                     case R.id.action_home:
                         fragment = new HomeFragment();
                         break;
-                    case R.id.action_compose:
-                        fragment = new ComposeFragment();
+                    case R.id.action_search:
+                        fragment = new SearchFragment();
                         break;
                     case R.id.action_profile:
                         fragment = new ProfileFragment();
