+++ app/src/main/java/com/example/photofy/fragments/SearchFragment.java
@@ -6,18 +6,15 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import com.example.photofy.R;
 
-public class LoadingFragment extends Fragment {
+public class SearchFragment extends Fragment {
 
-    public static final String TAG = "LoadingFragment";
-
-    public LoadingFragment() {
+    public SearchFragment() {
         // Required empty public constructor
     }
 
@@ -25,12 +22,11 @@ public class LoadingFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_loading, container, false);
+        return inflater.inflate(R.layout.fragment_search, container, false);
     }
 
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        Log.i(TAG, "loading");
     }
 }
