+++ app/src/main/res/layout/activity_loading.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.LoadingFragment">
+    tools:context=".activities.LoadingActivity">
 
     <com.airbnb.lottie.LottieAnimationView
         android:id="@+id/animationView"
