+++ app/src/main/res/menu/menu_compose.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/miOpenCamera"
+        android:title="@string/open_camera" />
+    <item
+        android:id="@+id/miUploadImage"
+        android:title="@string/upload_image" />
+</menu>
